+++ lib/karafka/web/pro/ui/routes/errors.rb
@@ -12,7 +12,7 @@ module Karafka
           class Errors < Base
             route do |r|
               r.on 'errors' do
-                controller = Controllers::ErrorsController.new(params)
+                controller = build(Controllers::ErrorsController)
 
                 r.get :partition_id, Integer do |partition_id, offset|
                   if params.current_offset != -1