+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - **[Feature]** Provide ability to manage topics configuration from the Web UI (Pro).
 - **[Feature]** Provide ability to manage topics partitioning from the Web UI (Pro).
 - **[Feature]** Provide ability to inject custom CSS and JS to adjust the Web UI.
+- [Enhancement] Optimize topic specific lookups (Pro).
+- [Enhancement] Replace simple in-process metadata cache with user tracking version for multi-process deployments improvements.
 - [Enhancement] Move web ui topics configuration into config.
 - [Enhancement] Upgrade DaisyUI to 5.0 and Tailwind to 4.0.
 - [Enhancement] Make consumer sampler/stats gathering compatible across debian/alpine/wolfi OSes (chen-anders)